requirements.txt
# dependencies for CyberEden (Host Layer)
pydantic>=2.0
python-dotenv>=1.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
google-genai>=1.0
google-api-core>=2.15.0
docker>=7.0.0
docker